@@ -4,13 +4,15 @@ version: 0.2.0
 homepage: https://github.com/wuzhendev/flutter_progress_dialog
 
 environment:
-  sdk: '>=2.12.0'
+  sdk: ^3.11.0
+  flutter: '>=3.41.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_platform_widgets: ^9.0.0
+  flutter_platform_widgets: ^10.0.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  flutter_lints: ^5.0.0
